==> 2_preprocess_making_token_ngram.ipynb <==
# 토큰화 및 N-gram화

1. 채권분석리포트

import pandas as pd
df=pd.read_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\채권분석리포트전처리.csv")

from ekonlpy.sentiment import MPCK
mpck = MPCK()

def make_ngram(df):
    try:
        tokens=mpck.tokenize(df['content'])
        ngrams=mpck.ngramize(tokens)
        tokens_text=",".join(tokens)
        ngrams_text=",".join(ngrams)
        total_text=tokens_text+","+ngrams_text
    except Exception as e:
        print(e)
        total_text=""
    return total_text

df['ngrams']=df.apply(make_ngram,axis=1)
df.to_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\total채권분석리포트.csv",index=False)

2. 네이버뉴스

import pandas as pd
df=pd.read_csv(r"C:\Users\HanBi\Downloads\뉴스전처리.csv",encoding='utf-8')

from ekonlpy.sentiment import MPCK
mpck = MPCK()

df['ngrams']=df.apply(make_ngram,axis=1)
