==> 2_preprocess_merging_dataframes.ipynb <==
# 데이터 취합

1. 네이버뉴스, 채권분석리포트, 금통위의사록 한 파일로 합치기 (columns:date/title/content/ngram)
2. date를 기준으로 콜금리파일과 merge

#네이버뉴스
df=pd.read_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\0108네이버뉴스.csv")
#채권분석리포트
df1=pd.read_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\total채권분석리포트.csv")
#금통위 의사록
df2=pd.read_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\minutes_data_save.csv")
total_df=pd.concat([df,df1,df2], axis=0)

#날짜로 재정렬
total_df1=total_df.sort_values(by=['date'], axis=0)

#콜금리
df_call=pd.read_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\콜금리데이터프레임.csv")
m_df=pd.merge(total_df1, df_call, on="date", how='left')

m_df=m_df.drop(['callrate','한달전 date', '한달전 callrate','Unnamed: 0'],axis=1)
m_df=m_df.dropna(axis=0)
m_df=m_df.drop_duplicates()


m_df.to_csv(r"C:\Users\HanBi\Desktop\2021\project1\crawling\final0108.csv", encoding='utf-8', index=False)
