==> 3_make_interest_dictionary.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("final0108.csv", sep=",")
df.dropna(inplace=True)  #결측치제거
df = df.reset_index(drop=True) #인덱스 초기화, 안하면 에러남

df_up = df[df['labeling'] == 'up']      #up 데이터 프레임과 down 데이터프레임 만들기
df_down = df[df['labeling']== 'down']

df_up

df_up = df_up.reset_index(drop=True)          #새로 만들어진 데이터 프레임의 인덱스 초기화
df_down = df_down.reset_index(drop=True)

ngram_list = []    #n그램을 나눠서 리스트화
for i in range(len(df_up)):
    ngram_list +=df_up['ngrams'][i].split(',')
    
    
ngram_list2 = []
for i in range(len(df_down)):
    ngram_list2 +=df_down['ngrams'][i].split(',')
    
    

up_ngram_df = pd.DataFrame({'count': ngram_list}) #위에서 만든 리스트를 데이터프레임으로
up_count_df = pd.DataFrame(up_ngram_df['count'].value_counts()).reset_index()      #value_counts함수를 이용해서 원하는결과 만들고 인덱스 초기화

down_ngram_df = pd.DataFrame({'count': ngram_list2}) 
down_count_df = pd.DataFrame(down_ngram_df['count'].value_counts()).reset_index()

up_count_df.rename(columns = {'index':'ngram','count':'up_count'},inplace=True)  #컬럼명변경
down_count_df.rename(columns={'index':'ngram','count':'down_count'},inplace=True)

df1 =pd.merge(up_count_df,down_count_df, how='outer', on='ngram')         #
df2=df1.fillna(0) #결측치를 0으로



df2['total_count'] = (df2['up_count']+df2['down_count'])  #15미만의 자료 제거하기 위한total_count 구하기 
df2

df3 = df2[df2['total_count'] >=15]        #15미만의 자료 제거
df3

print(df3['up_count'].sum(axis=0))
print(df3['down_count'].sum(axis=0))

#Polarity Score 구하기
df3['polar_score'] = ((df3['up_count'] / sum(df3['up_count']) )/(df3['down_count']/sum(df3['down_count'])))

df3

def polarity(n):                                    #폴라리티 구분 함수
    if n > 1.3:
        return 'Hawkish'
    elif n < 0.76:
        return 'Dovish'
    else:
        return 'nothing'
    

df3['polarity'] = df3['polar_score'].apply(polarity)                  #apply함수를 이용하요 폴라리티를 구하기

df3

df_Neg = df3[df3['polarity'] == 'Dovish'] #부정사전 긍정사전 만들기
df_Pos = df3[df3['polarity'] == 'Hawkish']

df_Neg=df_Neg.sort_values('total_count',ascending=False) #정렬하기
df_Pos=df_Pos.sort_values('total_count',ascending=False)

df_Pos=df_Pos.reset_index(drop=True)             #인덱스 초기화
df_Neg=df_Neg.reset_index(drop=True)

df_Pos.head(20) #사전 완성도를 위하여 상위 20개를 뽑아봄

Dict_pos = df_Pos.drop([1,2,3,12])                                  #긍정사전 최종완성
Dict_pos=Dict_pos.sort_values('up_count',ascending=False)
Dict_pos=Dict_pos.reset_index(drop=True)
Dict_pos

Dict_neg = df_Neg.drop([11,12])                               #부정사전 최종완성
Dict_neg=Dict_neg.sort_values('down_count',ascending=False)
Dict_neg=Dict_neg.reset_index(drop=True)
Dict_neg

Dict_pos.to_csv(r"C:\Users\kyt50\anaconda3\kdigital\project\긍정사전.csv", encoding='utf-8', index=False)     #저장하기
Dict_neg.to_csv(r"C:\Users\kyt50\anaconda3\kdigital\project\부정사전.csv", encoding='utf-8', index=False)
